double right click is disabled
This file will make all maps listed in the file to be able to zoom out even farther. However, double right click (reset to default camera angle) is disabled for those maps. In order to reverse this file, just delete everything (or just the maps you want) in the text file.